rewrite the entire freaking bot to use strife.js
forgive if buggers :sob:
