+++ package.json
@@ -23,9 +23,10 @@
 		"@types/axios": "0.14.0",
 		"axios": "0.27.2",
 		"discord-api-types": "0.37.2",
-		"discord.js": "13.6.0",
+		"discord.js": "^14.15.2",
 		"dotenv": "16.0.1",
-		"mongoose": "6.5.3"
+		"mongoose": "6.5.3",
+		"strife.js": "^4.0.1"
 	},
 	"devDependencies": {
 		"@redguy12/prettier-config": "1.0.0",
